Rename setup to dev
