--- .vscode/launch.json
@@ -1,21 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Debug locally with SQLite",
-      "skipFiles": [
-        "<node_internals>/**"
-      ],
-      "program": "${workspaceFolder}\\setup\\runner.js",
-      "env": {
-        "LOCAL_SSCCE": "true",
-        "DIALECT": "sqlite"
-      }
-    }
-  ]
-}
