+++ dev/db2/latest/reset.sh
