+++ dev/db2/latest/stop.sh
