+++ dev/db2/oldest/reset.sh
