+++ dev/mariadb/latest/reset.sh
