+++ dev/mysql/latest/reset.sh
