+++ dev/sequelize-version.cjs
