+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "include": [
     "src/*",
-    "setup/*"
+    "dev/*"
   ],
   "compilerOptions": {
     "target": "es2021",
